requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
murmurhash==1.0.15
networkx==3.6.1
numpy==2.4.2
openai==2.24.0
packaging==26.0
pillow==12.1.1
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
rich==14.3.3
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
thinc==8.3.10
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.2
